tests/wpt/tests/encoding-detection/it-windows-1252-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>it windows-1252 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1252", 'Expected windows-1252');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/it-windows-1252-late.sub.html");
</script>
